:sparkles: Start fleshing out the main screen
Still a bunch of stuff to do here, but this is the core idea I have for
this: I want a list of keys we expect and which need to be pressed and
tested, and a list of keys that have been tested. Testing should happen only
in a specific widget at the top of the screen.

Things to come will include a log of names we didn't expect, and hopefully a
log of sequences that didn't record well (so perhaps a sequence of keys that
start with an escape and then flow through really fast) and which the user
can annotate.

I also want to allow adding notes to triggered keys so the user can
highlight if a key was seemingly triggered incorrectly (so if one key is
pressed but another name came through).